Fix layouts, createRecipe and account deletion
New user are now assigned layouts of empty objects to avoid null.

When deleting an account, the user's recipes and lists are first deleted.

Function createRecipe is wrapped in useCallback with targetPage as dependencies.
Otherwise targetPage is updated but not in createRecipe, so even when page changes
from list to recipes, targetPage within createRecipe will still equal "lists".